+++ module/LearnZF2I18n/languages/phpArray/en_US.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'Hello' => 'Hello',
+    'Learn Zend Framework with this tutorials' => 'Learn Zend Framework with this tutorials',
+    'Change current language' => 'Change current language',
+];