+++ module/LearnZF2I18n/languages/phpArray/id_ID.php
@@ -1,0 +1,9 @@
+<?php
+
+// Indonesian Bahasa
+
+return [
+    'Hello' => 'Halo',
+    'Learn Zend Framework with this tutorials' => 'Belajar Zend Framework melalui tutorial-tutorial ini',
+    'Change current language' => 'Ganti bahasa saat ini',
+];